Global error handler & Not found middleware added

+++ src/modules/Comments/comments.services.ts
@@ -14,6 +14,12 @@ const createComment = async (payLoad: {
 };
 
 const getAllComments = async () => {
+  const isExist = await prisma.comments.findFirst();
+
+  if (!isExist) {
+    throw new Error("NO_COMMENT_EXIST_YET");
+  }
+
   const result = await prisma.comments.findMany({
     orderBy: {
       created_At: "asc",
@@ -38,10 +44,25 @@ const getCommentById = async (id: string) => {
       },
     },
   });
+
+  if (!result) {
+    throw new Error("COMMENT_NOT_FOUND");
+  }
+
   return result;
 };
 
 const getCommentByAuthorId = async (author_Id: string) => {
+  const isExist = await prisma.comments.findFirst({
+    where: {
+      author_Id,
+    },
+  });
+
+  if (!isExist) {
+    throw new Error("NO_COMMENT_CREATED_BY_THE_AUTHOR / INVALID_AUTHOR_ID");
+  }
+
   const result = await prisma.comments.findMany({
     where: {
       author_Id,
@@ -69,7 +90,7 @@ const deleteComment = async (
   author_Id: string,
   isAdmin: boolean
 ) => {
-  const isExist = await prisma.comments.findUniqueOrThrow({
+  const isExist = await prisma.comments.findUnique({
     where: {
       id,
     },
@@ -79,8 +100,12 @@ const deleteComment = async (
     },
   });
 
+  if (!isExist) {
+    throw new Error("NO_SUCH_COMMENT_EXIST_TO_DELETE");
+  }
+
   if (!isAdmin && author_Id !== isExist.author_Id) {
-    throw new Error("You are not allowed to perform this action");
+    throw new Error("YOU_ARE_NOT_ALLOWED_TO_PERFORM_THIS_ACTION");
   }
 
   const result = await prisma.comments.delete({
@@ -97,7 +122,7 @@ const updateComment = async (
   author_Id: string,
   isAdmin: boolean
 ) => {
-  const isExist = await prisma.comments.findUniqueOrThrow({
+  const isExist = await prisma.comments.findUnique({
     where: {
       id,
     },
@@ -108,18 +133,22 @@ const updateComment = async (
     },
   });
 
+  if (!isExist) {
+    throw new Error("NO_SUCH_COMMANT_EXIST_TO_UPDATE");
+  }
+
   const updateData = { ...payLoad };
 
   if (!isAdmin) {
     if (author_Id !== isExist.author_Id) {
-      throw new Error("You are not allowed to perform this action");
+      throw new Error("YOU_ARE_NOT_ALLOWED_TO_PERFORM_THIS_ACTION");
     }
 
     if (updateData.status && updateData.status === isExist.status) {
       delete updateData.status;
 
       if (Object.keys(updateData).length === 0) {
-        throw new Error(`status : ${isExist.status} already exists.`);
+        throw new Error(`STATUS : ${isExist.status} ALREADY EXIST.`);
       }
     }
   }
@@ -131,13 +160,13 @@ const updateComment = async (
       delete updateData.status;
 
       if (Object.keys(updateData).length === 0) {
-        throw new Error(`status : ${isExist.status} already exists.`);
+        throw new Error(`STATUS : ${isExist.status} ALREADY EXIST.`);
       }
     }
 
     if (Object.keys(updateData).length === 0) {
       throw new Error(
-        "No valid fields provided for update.You can only update the post status."
+        "NO_VALID_FIELD_PROVIDED_FOR_UPDATE.YOU_CAN_ONLY_UPDATE_POST_STATUS."
       );
     }
   }
